Removed (long) from the square equation

+++ c_version/kaprekarUtils.c
@@ -30,7 +30,7 @@ int isKaprekar(int n) {
 
   int i;
   //compute the square
-  long square = n * (long) n;
+  long square = n * n;
   //compute the number of digits of the square
   int numDigits = (int) log10(n) + 1;
   long modulus = 0;
